#0: code refactoring
