+++ .github/workflows/fast-dispatch-full-regressions-and-models-impl.yaml
@@ -88,7 +88,7 @@ jobs:
       fail-fast: false
       matrix:
         card: [N150, N300]
-        model: [common_models, functional_unet, llama3.2-1B, qwen, mistral7b, mistral7b_eth, resnet50, yolov4, whisper, yolov8s_world, stable_diffusion_xl_base]
+        model: [common_models, functional_unet, llama3.2-1B, qwen, mistral7b, mistral7b_eth, resnet50, yolov4, whisper, yolov8s_world, stable_diffusion_xl_base, ufld_v2]
     name: Nightly ${{ matrix.card }} ${{ matrix.model }}
     defaults:
       run:
