+++ tests/nightly/single_card/ufld_v2/test_ufld_v2_modules.py
@@ -1,0 +1,1 @@
+../../../ttnn/integration_tests/UFLD_v2/test_ttnn_UFLD_v2.py
